# public/openapi.json
{
  "openapi": "3.1.0",
  "info": { "title": "wordlegpt API", "version": "1.0" },
  "components": {
    "schemas": {}
  },
  "security": [],
  "servers": [{ "url": "https://wordlegpt.vercel.app" }],
  "paths": {
    "/api/answers": {
      "get": {
        "summary": "Returns the number of past Wordle answers",
        "description": "Gets past Wordle answers from the Rock Paper Shotgun website and returns the number of past answers.",
        "operationId": "getAnswersLength",
        "responses": {
          "200": {
            "description": "A JSON object containing the length of answers.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "answersLength": {
                      "type": "integer",
                      "description": "The number of past Wordle answers.",
                      "example": 1462
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/guesses/check": {
      "get": {
        "summary": "Check if a word is a valid guess and if it was a past answer",
        "description": "Checks from the list of valid wordle words and past answers from the Rock Paper Shotgun website if the word is a valid guess and if it was a past answer.",
        "operationId": "checkWord",
        "parameters": [
          {
            "in": "query",
            "name": "word",
            "schema": { "type": "string" },
            "required": true,
            "description": "The word to check"
          }
        ],
        "responses": {
          "200": {
            "description": "Indicates if the word is a valid guess and if it was a past answer",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "isPastAnswer": {
                      "type": "boolean",
                      "description": "Indicates if the word was a past answer"
                    },
                    "isValidGuess": {
                      "type": "boolean",
                      "description": "Indicates if the word is a valid guess"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid word provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/guesses/match": {
      "get": {
        "summary": "Retrieve matching guesses based on the provided pattern and options.",
        "operationId": "getMatchingGuesses",
        "parameters": [
          {
            "in": "query",
            "name": "pattern",
            "schema": { "type": "string" },
            "required": true,
            "description": "A 5-character pattern where letters represent fixed positions and asterisks (*) represent wildcards."
          },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 25 },
            "required": false,
            "description": "The maximum number of matches to return."
          },
          {
            "in": "query",
            "name": "exclude",
            "schema": { "type": "string" },
            "required": false,
            "description": "A list of letters that cannot be in the word"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of matching guesses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "matches": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/guesses": {
      "get": {
        "description": "Returns the number of guesses and potentially winning guesses",
        "operationId": "getGuessesLength",
        "responses": {
          "200": {
            "description": "A JSON object containing the length of guesses and potentially winning guesses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "guessesLength": {
                      "type": "integer",
                      "description": "The total number of guesses"
                    },
                    "potentiallyWinningGuessesLength": {
                      "type": "integer",
                      "description": "The number of potentially winning guesses"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/guesses/shoot": {
      "get": {
        "summary": "Generates random guesses",
        "description": "Returns a list of random guesses that are valid answers but not in the answers list from the Rock Paper Shotgun website.",
        "operationId": "getGuesses",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 1, "minimum": 1 },
            "description": "The number of random guesses to return."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of random guesses",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "guesses": {
                      "type": "array",
                      "items": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request parameters",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "Error message"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "tags": []
}